requirements.txt
# LLM API Clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# CLI and Terminal
click>=8.0.0
rich>=13.0.0

# PDF Report Generation
reportlab>=4.0.0

# Data Processing
numpy>=1.24.0
scipy>=1.10.0

# Utilities
python-dateutil>=2.8.0

# Local LLM Support (optional - uncomment to enable)
# llama-cpp-python>=0.2.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
